pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.7"
dependencies = [
    "answerrocket-client",
    "mcp[cli]>=1.10.1",
]

[tool.uv.sources]
answerrocket-client = { git = "ssh://git@github.com/answerrocket/answerrocket-python-client.git", rev = "get-copilots-for-mcp" }
